Added a Context Manager for Acquiring a File-Level Lock
If multiple processes want to acquire the same file, they should use this context manager to acquire a file lock first. Once the lock is acquired, control is yielded back to the calling program. This was tested with 100 concurrent threads and processes to work out any bugs, but this is not guaranteed to be bug-free yet.